requirements.txt
jax
diffrax
numpy
matplotlib
jupyter
ipympl
ipywidgets
absl-py
ml-collections
flax
clu
jaxtyping
tensorflow
datasets
orbax-checkpoint